Feat: implement workbook detail page API using auth fetch part of SWM-153 related to #52
