+++ src/app/(after-login)/(top)/workbookDetail/workbook-detail-api.ts
@@ -1,92 +1,8 @@
-import { IWorkbookBase } from '@/model/i-workbook';
 import { IQuestion } from '@/model/i-question';
 
 let session: any;
 if (typeof window !== 'undefined') {
   session = localStorage.getItem('accessToken');
-}
-
-export async function readWorkbookById(workbookId: string) {
-  try {
-    const response = await fetch(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/workbook/${workbookId}`,
-      {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${session}`,
-        },
-      },
-    );
-
-    return await response.json();
-  } catch (e) {
-    console.error(`readWorkbookById failed by ${e}`);
-  }
-}
-
-export async function readQuestionsByWorkbookId(workbookId: string) {
-  try {
-    const response = await fetch(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/workbook/questions/${workbookId}`,
-      {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${session}`,
-        },
-      },
-    );
-
-    return await response.json();
-  } catch (e) {
-    console.error(`readQuestionsByWorkbookId failed by ${e}`);
-  }
-}
-
-export async function updateWorkbookDetail({
-  workbookBase,
-  workbookId,
-}: {
-  workbookBase: IWorkbookBase;
-  workbookId: string;
-}) {
-  try {
-    const response = await fetch(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/workbook/${workbookId}`,
-      {
-        method: 'PUT',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${session}`,
-        },
-        body: JSON.stringify(workbookBase),
-      },
-    );
-
-    return await response.json();
-  } catch (e) {
-    console.error(`updateWorkbookDetail failed by ${e}`);
-  }
-}
-
-export async function deleteWorkbook(workbookId: string) {
-  try {
-    const response = await fetch(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/workbook/${workbookId}`,
-      {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${session}`,
-        },
-      },
-    );
-
-    return await response.json();
-  } catch (e) {
-    console.error(`deleteWorkbook failed by ${e}`);
-  }
 }
 
 export async function updateQuestionSequence({
